Fix ragged masking out of bounds error.
Some meshes (e.g. QU240km) don't do masking of ragged indices
with 0. Instead they use `nInidices+1` as the invalid value,
which can produce out of bounds errors. Check if that the case,
and if so set the out of bounds indices to (-1).

Also moved the zero indexing to the minimal dataset construction
to ensure it's done and done uniformly across arrays.