Ensure vertex patches encompass the full kite area
If both an edge and cell are missing at an edge node, collpase
to the first node (not the vertex) to ensure patch encompasses
the full kite area. If both the edge and cell are missing the
next (cell) node will be collpased to the vertex, and because
the edges lead the cells this will work.